+++ tests/Unit/DependencyInjection/ExtensionContractTest.php
@@ -77,6 +77,8 @@ use Oronts\AssetPilotBundle\Service\ConfigValidator;
 use Oronts\AssetPilotBundle\Service\ConfigValidatorInterface;
 use Oronts\AssetPilotBundle\Service\ContentUsageScanner;
 use Oronts\AssetPilotBundle\Service\ContentUsageScannerInterface;
+use Oronts\AssetPilotBundle\Service\Conversion\AssetConverterResolver;
+use Oronts\AssetPilotBundle\Service\Conversion\AssetConverterResolverInterface;
 use Oronts\AssetPilotBundle\Service\DbalApplyPlanClaimStore;
 use Oronts\AssetPilotBundle\Service\DbalDependencyProjection;
 use Oronts\AssetPilotBundle\Service\DbalDependencyProjectionFreshness;
@@ -437,6 +439,7 @@ final class ExtensionContractTest extends TestCase
             MovePlannerInterface::class => MovePlanner::class,
             OrganizeDispatcherInterface::class => OrganizeDispatcher::class,
             AutomaticOrganizeIntentStoreInterface::class => AutomaticOrganizeIntentStore::class,
+            AssetConverterResolverInterface::class => AssetConverterResolver::class,
             AssetZipServiceInterface::class => AssetZipService::class,
             ApplyPlanServiceInterface::class => ApplyPlanService::class,
             RulePreviewPlanServiceInterface::class => RulePreviewPlanService::class,